Remove PortfolioEngine implementation from engine_v2.py, consolidating portfolio management functionality.
